requirements.txt
fastapi
uvicorn[standard]
celery
redis
requests
pydantic[email]
pwdlib
pyjwt
